+++ internal/commands/.scripts/integration_up.sh
@@ -1,3 +1,13 @@
+#!/bin/bash
+
+# Check if the first argument is provided
+if [ -z "$1" ]; then
+  echo "Usage: $0 <test-folder>"
+  exit 1
+fi
+
+TEST_FOLDER=$1
+
 docker run \
   --name squid \
   -d \
@@ -10,13 +20,23 @@ wget https://sca-downloads.s3.amazonaws.com/cli/latest/ScaResolver-linux64.tar.g
 tar -xzvf ScaResolver-linux64.tar.gz -C /tmp
 rm -rf ScaResolver-linux64.tar.gz
 
-go test \
-  -tags integration \
-  -v \
-  -timeout 210m \
-  -coverpkg github.com/checkmarx/ast-cli/internal/commands,github.com/checkmarx/ast-cli/internal/services,github.com/checkmarx/ast-cli/internal/wrappers \
-  -coverprofile cover.out \
-  github.com/checkmarx/ast-cli/test/integration
+if [ "$TEST_FOLDER" == "projectstest" ]; then
+  go test \
+    -tags integration \
+    -v \
+    -timeout 210m \
+    -coverpkg github.com/checkmarx/ast-cli/internal/commands,github.com/checkmarx/ast-cli/internal/services/projects.go,github.com/checkmarx/ast-cli/internal/wrappers/projects.go \
+    -coverprofile cover.out \
+    github.com/checkmarx/ast-cli/test/integration/projectstest
+else
+    go test \
+      -tags integration \
+      -v \
+      -timeout 210m \
+      -coverpkg github.com/checkmarx/ast-cli/internal/commands,github.com/checkmarx/ast-cli/internal/services,github.com/checkmarx/ast-cli/internal/wrappers \
+      -coverprofile cover.out \
+      github.com/checkmarx/ast-cli/test/integration
+fi
 
 status=$?
 echo "status value after tests $status"
